// utils/machines/characters/enums.ts
export enum STATES {
  ON_CHARACTERS = 'ON_CHARACTERS',
  GET_CHARACTERS = 'GET_CHARACTERS',
  ERROR = 'ERROR',
  SUCCESS = 'SUCCESS'
}

export enum EVENTS {
  NEXT_STEP = 'NEXT_STEP',
  PREVIOUS_STEP = 'PREVIOUS_STEP',
  GET_CHARACTERS = 'GET_CHARACTERS',

  RESTART = 'RESTART'
}

export enum SERVICES {
  GET_CHARACTERS = 'GET_CHARACTERS'
}